requirements.txt
python-telegram-bot==20.3
huggingface-hub==0.16.4
Pillow==9.5.0
